requirements.txt
asgiref==3.8.1
Django==4.2.13
sqlparse==0.5.0
typing_extensions==4.12.2
psycopg[binary]==3.1.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
debugpy==1.7.0
asgiref==3.8.1
boto3==1.34.131
botocore==1.34.131
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.2
pillow==10.3.0
python-magic==0.4.27
django-filter==24.2
django-cors-headers==4.4.0